requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pypdf==4.3.1
reportlab==4.2.2
Pillow==10.4.0
